Conflicts:
	admin/templates/bootstrap/tpl_inc/bestellungen_uebersicht.tpl

+++ dbeS/Artikel_xml.php
@@ -518,7 +518,11 @@ function bearbeiteInsert($xml, array $conf)
             DBUpdateInsert('tartikelkonfiggruppe', $oArtikelKonfig_arr, 'kArtikel', 'kKonfiggruppe');
         }
         // Sonderpreise
-        Shop::DB()->delete('tsonderpreise', 'kArtikelSonderpreis', $artikel_arr[0]->kArtikel);
+        Shop::DB()->query("
+            DELETE asp, sp
+                FROM tartikelsonderpreis asp LEFT JOIN tsonderpreis sp ON sp.kArtikelSonderpreis = asp.kArtikelSonderpreis
+                WHERE asp.kArtikel = " . $artikel_arr[0]->kArtikel,
+            4);
         if (isset($xml['tartikel']['tartikelsonderpreis'])) {
             updateXMLinDB($xml['tartikel']['tartikelsonderpreis'], 'tsonderpreise', $GLOBALS['mSonderpreise'], 'kArtikelSonderpreis', 'kKundengruppe');
         }
@@ -1107,45 +1111,48 @@ function loescheStueckliste($kStueckliste)
 function fuelleKategorieGesamt($oKategorieArtikel_arr)
 {
     if (is_array($oKategorieArtikel_arr) && count($oKategorieArtikel_arr) > 0) {
-        $isPageCacheEnabled = Shop::Cache()->isPageCacheEnabled();
-        // Laufe jeden KategorieArtikel Eintrag durch
+        $deleted = array();
+        $added   = array();
+        //$oKategorieArtikel_arr probably always contains the same kArtikel. this is just to be sure.
+        foreach ($oKategorieArtikel_arr as $oKategorieArtikel) {
+            $kArtikel = (int)$oKategorieArtikel->kArtikel;
+            if (!in_array($kArtikel, $deleted)) {
+                $deleted[] = $kArtikel;
+                Shop::DB()->delete('tkategorieartikelgesamt', 'kArtikel', (int)$oKategorieArtikel->kArtikel);
+            }
+        }
         foreach ($oKategorieArtikel_arr as $oKategorieArtikel) {
             // Lösche aktuellen KategorieArtikel
-            Shop::DB()->delete('tkategorieartikelgesamt', 'kArtikel', (int) $oKategorieArtikel->kArtikel);
             $oOberKategorie_arr = array();
             // Hole die Kategorie vom aktuellen KategorieArtikel
-            $oKategorie = Shop::DB()->query("SELECT * FROM tkategorie WHERE kKategorie = " . (int) $oKategorieArtikel->kKategorie, 1);
-
+            $oKategorie = Shop::DB()->select('tkategorie', 'kKategorie', (int)$oKategorieArtikel->kKategorie);
             $oOberKategorie_arr[] = $oKategorie;
-            Shop::Cache()->flushTags(array(CACHING_GROUP_CATEGORY . '_' . (int) $oKategorieArtikel->kKategorie));
+            Shop::Cache()->flushTags(array(CACHING_GROUP_CATEGORY . '_' . (int)$oKategorieArtikel->kKategorie));
             // Laufe solange bis es keine OberKategorie mehr zum aktuellen KategorieArtikel gibt
             // Falls es zum aktuellen KategorieArtikel keine OberKategorie gibt, wird die schleife nicht betreten
             while (isset($oKategorie->kOberKategorie) && $oKategorie->kOberKategorie > 0) {
                 // Hole OberKategorie
-                $oKategorie = Shop::DB()->query("SELECT * FROM tkategorie WHERE kKategorie = " . (int) $oKategorie->kOberKategorie, 1);
+                $oKategorie = Shop::DB()->select('tkategorie', 'kKategorie', (int)$oKategorie->kOberKategorie);
                 if (isset($oKategorie->kKategorie)) {
                     $oOberKategorie_arr[] = $oKategorie;
                 }
             }
 
-            $oOberKategorie_arr = array_reverse($oOberKategorie_arr);  // Dreh das Array um, damit wir an Array[0] auch das Level 0 haben
+            $oOberKategorie_arr = array_reverse($oOberKategorie_arr); // Dreh das Array um, damit wir an Array[0] auch das Level 0 haben
 
             if (count($oOberKategorie_arr) > 0) {
                 // Speicher den kompletten Kategoriepfad zum aktuellen KategorieArtikel nach Level sortiert in die Datenbank
                 foreach ($oOberKategorie_arr as $i => $oOberKategorie) {
-                    $oKategorieArtikelGesamt                 = new stdClass();
-                    $oKategorieArtikelGesamt->kArtikel       = (int)$oKategorieArtikel->kArtikel;
-                    $oKategorieArtikelGesamt->kOberKategorie = (int)$oOberKategorie->kOberKategorie;
-                    $oKategorieArtikelGesamt->kKategorie     = (int)$oOberKategorie->kKategorie;
-                    $oKategorieArtikelGesamt->nLevel         = $i;
+                    if (!in_array((int)$oOberKategorie->kKategorie, $added)) {
+                        $oKategorieArtikelGesamt                 = new stdClass();
+                        $oKategorieArtikelGesamt->kArtikel       = (int)$oKategorieArtikel->kArtikel;
+                        $oKategorieArtikelGesamt->kOberKategorie = (int)$oOberKategorie->kOberKategorie;
+                        $oKategorieArtikelGesamt->kKategorie     = (int)$oOberKategorie->kKategorie;
+                        $oKategorieArtikelGesamt->nLevel         = $i;
 
-                    Shop::DB()->insert('tkategorieartikelgesamt', $oKategorieArtikelGesamt);
-                    Shop::Cache()->flushTags(array(CACHING_GROUP_CATEGORY . '_' . $oKategorieArtikelGesamt->kKategorie));
-                    if ($isPageCacheEnabled === true) {
-                        if (!isset($smarty)) {
-                            $smarty = Shop::Smarty();
-                        }
-                        $smarty->clearCache(null, 'jtlc|category|cid' . $oKategorieArtikelGesamt->kKategorie);
+                        Shop::DB()->insert('tkategorieartikelgesamt', $oKategorieArtikelGesamt);
+                        Shop::Cache()->flushTags(array(CACHING_GROUP_CATEGORY . '_' . $oKategorieArtikelGesamt->kKategorie));
+                        $added[] = (int)$oOberKategorie->kKategorie;
                     }
                 }
             }
